Bring back the bridges per distributor metrics
The right place to do that will be rdsys, but let's keep the metrics as
well in bridgedb so the migration for the metrics team is more smooth.
With the migration to rdsys we loose the metrics of the unasigned
bridges and any distributor not managed by bridgedb.

Related: rdsys#67